Change int aggregation to double aggregation

+++ aggregation/src/main/java/com/tempest/aggregation/CollectingAggregator.java
@@ -1,10 +1,9 @@
 package com.tempest.aggregation;
 
 import com.tempest.aggregation.pojo.AggregationKey;
-import com.tempest.metric.MetricEvent;
 
 import java.util.Map;
 
 public interface CollectingAggregator extends MetricEventConsumer {
-    Map<AggregationKey, Integer> collectAndReset(); // Returns current aggregated counts and clears state
+    Map<AggregationKey, Double> collectAndReset(); // Returns current aggregated counts and clears state
 }
